% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om.R
\name{om}
\alias{om}
\title{Run operating model}
\usage{
om(a = NULL, data_om, seed = NULL, verbose = TRUE)
}
\arguments{
\item{a}{\code{NULL} or the number of time periods to run OM before updating management plan}

\item{data_om}{tagged list containing model parameters}

\item{seed}{seed for random number generation}

\item{verbose}{if \code{TRUE} then runtime feedback is provided}
}
\value{

}
\description{
\code{om} runs the \code{operating model}
}
