% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{plot_map}
\alias{plot_map}
\title{EXPERIMENTAL}
\usage{
plot_map(
  z_s,
  inla_proj,
  land,
  bath,
  legend_type = 1,
  p,
  y,
  leglim = c(min(z_s), max(z_s)),
  xlim,
  ylim,
  np = 12,
  ncols = 100,
  pretty = TRUE,
  cex.text = 1,
  col.text = "black",
  bg = "dark blue",
  ...
)
}
\description{
EXPERIMENTAL
}
