% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{plot_map_cy}
\alias{plot_map_cy}
\title{EXPERIMENTAL}
\usage{
plot_map_cy(
  z_csy,
  inla_proj,
  land,
  bath,
  lab_y,
  p,
  y,
  leglim_c = NULL,
  cex.text = 1,
  xlim,
  ylim,
  col.text = "black",
  no = 1,
  lab_om = 1:no,
  np = 12,
  ncols = 100,
  pretty = TRUE,
  species_names = NULL,
  height = 5,
  p_width = 0.3,
  mar = rep(0.2, 4),
  legend_type = 1,
  legend_size = 0.4,
  bg = "dark blue",
  widthhint = 0.55,
  padin_arr = c(0, 0),
  padin_bar = c(0, 0)
)
}
\description{
EXPERIMENTAL
}
