% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{plot_map_gif}
\alias{plot_map_gif}
\title{EXPERIMENTAL}
\usage{
plot_map_gif(
  z_i_st,
  inla_proj,
  land,
  bath,
  p1 = 1,
  y1 = 2018,
  leglim_i = lapply(z_i_st, function(z_st) c(0, max(z_st))),
  nt = ncol(z_i_st[[1]]),
  np = 12,
  ncols = 100,
  filename,
  fps = 10,
  ...
)
}
\description{
EXPERIMENTAL
}
